==> 1.Data_wrangling_and_cleaning.ipynb <==
# Wrangling Twitter data
In this notebook, we are going to start out with the raw Twitter data that was captured using the Tweepy tool on which you can read more about in [this notebook](0.Tweepy_Bot.ipynb). The data captured was restricted to tweets from Alberta from February 25th  2019 to March 1st 2019. 

In order to prep the data for analysis, we will "wrangle" the data to clean it and structure it into a more useful format. 

## Goal: 
Learn some common aspects of data wrangling and convert the data set into a format more practical for further data exploration and analysis. 


## Data wrangling introduction
[Data wrangling](https://en.wikipedia.org/wiki/Data_wrangling) "is the process of transforming and mapping data from one "raw" data form into another format with the intent of making it more appropriate and valuable for a variety of downstream purposes such as analytics" (Wikipedia).

The purpose of data wrangling notebook is to transform complex data into a format that is easier to work with in order to focus on the analysis of data, rather than the wrangling in the next notebooks. 

#### Common data wrangling steps:
 - **Discovering**  - understanding the data
 - **Structuring** -  organizing the data
     - Naming or renaming variables
     - Changing the type of variables
     - Creating new variables through functions of existing of variables
     - Conditional processing of variables (i.e the values of new variable is based on the values of existing variables) 
 - **Cleaning** 
     - Redesign the data into a usable/functional format and correct/remove any bad data
     - Imputing missing values in a variables
 - **Publishing**
    - Selecting columns from input dataset to output dataset


## Load python modules

There are many packages available in Python that provide a variety of functions for data science. Here we load the libraries needed for this notebook. 

One of the most common libraries used for data science is [pandas](https://pandas.pydata.org/). There are good tutorials available on pandas and worthwhile spending more time with in order to become more comfortable with data science in python. 

import urllib.request
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline

import ast

## Step 1. Taking a look at the raw data


#### Reading the Twitter data  and displaying the first 5 rows

The raw data(`tweets.txt`) is stored in the same directory, we will read it into a pandas DataFrame. A **DataFrame** is a commonly encountered way to work with data in data science. It can be thought of as a 2D data structure that makes it easy to work with the data at hand. 

#target_url="https://swift-yeg.cloud.cybera.ca:8080/v1/AUTH_233e84cd313945c992b4b585f7b9125d/geeky-summit/alberta_tweets_feb_25_start.txt"
file_name="../data-science-for-entrepreneurs-helper/alberta_tweets_feb_25_start.txt"

## comment out this line to use your local copy 
#urllib.request.urlretrieve(target_url, file_name)

tweets = []
with open(file_name, 'r') as f:
    for lines in f.readlines():
        tweets.append(ast.literal_eval(lines))
tweets  = pd.DataFrame(tweets)

tweets.head(5) # head is a convenient way for showing the first few rows of a dataframe

#### Understanding the raw data

Let's take a look at some basic features of the raw data. For example, it's always a good idea to look at the columnns and the size of the data in question. 

tweets.columns

tweets.shape

There are several interesting features there, such as latitude and longitude of the tweets (`coordinates` column), and most of the columns make sense and are pretty intuitive. Two columns that are a bit confusing are `text` and `extended_tweet` as it looks like they may contain the same content. We should figure out if they are actually different. 

#### Indexing in DataFrames

DataFrames are convenient for working with data and are popular both in Python and R. There are a few different ways for accessing data inside a DataFrame. We'll cover a couple of aspects here, but you may also want to read more in the  pandas documentation or cheat sheet such as the one handed out [here](https://www.dataquest.io/blog/large_files/pandas-cheat-sheet.pdf). 

pd.set_option('display.max_row', 10) #Ensure only 10 rows are displayed to save space
pd.set_option('display.max_colwidth', -1) #Ensure output is not truncated by width

#Selecting a column
tweets[["text"]] #Using two brackets displays the column as a dataframe

#Selecting two columns 
tweets[["text", "extended_tweet"]]

#Selecting row subsets
selectone = tweets[["text", "extended_tweet"]].iloc[[2]] #iloc allows for selection by integers
selecttwo = tweets[["text", "extended_tweet"]].iloc[[2,5]] 
selectrange = tweets[["text", "extended_tweet"]].iloc[2:5]

print(selectone)
print(selecttwo)
print(selectrange)

#Select a certain row, all columns
tweets.iloc[3,:]

#### Investigating `text` and `extended_tweet` graphically

One way to investigate the difference between `text` and `extended_tweet` is to plot the length distribution for both columns. To do this, we use the [matplotlib library](https://matplotlib.org/).

#Set up a plot with subplots 
plt.subplot(1, 2, 1) # (rows, columns, active plot)
#Create a histogram, which counts the number of tweets of each length
plt.hist(tweets["text"].apply(len).values, color = 'green', edgecolor = 'black')
#Label the plot
plt.title('text')
plt.xlabel('Tweet length')
plt.ylabel('Number of tweets')

# Do the same thing for the second subplot
plt.subplot(1, 2, 2)
plt.hist(tweets["extended_tweet"].apply(len).values, color = 'blue', edgecolor = 'black')
plt.tight_layout()  #automatically adjusts subplot params so that the subplot(s) fits in to the figure area
plt.title('extended_tweet')
plt.xlabel('Tweet length')

Looks like `extended_tweet` has much longer tweets, up to 500(or more) characters in length, whereas there are no tweets longer than 150 characters in `text`. 

What does the longest tweet look like? To do this, we can easily pull the longest tweet from the DataFrame. 

print("Maximum length for extended_tweet column: ",max(tweets["extended_tweet"].apply(len)), "\n")

print("This is the row number for the longest tweet:")
print([tweets["extended_tweet"].apply(len).idxmax(axis=1)], "\n")

#### Exercise1: Using the row number you determined above, add your code to display what the longest tweet looks like in the `text` and `extended_tweet` columns in the cell below.

print("Tweet as seen in the **extended_tweet** column:\n")
###### Enter your code in place of None below
extended = None
######
print(str(extended) + "\n")

print("Tweet as seen in the **text** column:\n")
###### Enter your code in place of None below
text_tweet = None
######
print(text_tweet)

The answer should show that the `text` column is the truncated version of `extended_tweet` column.

## Step 2. (Re-) structuring  the data

One aspect to always pay close attention to is the data type of each column. Overlooking data types and ending up with unexpected types often leads to issues further downstream in the analysis. Here we will go through and change some of the data types to make sure they are what we want them to be. 

Examine DtataFrame column types/values first, this code snippet will print:
* column name
* column type
* sample value in the column

for column in tweets.columns:
    value=tweets[tweets[column].notnull()][column].iloc[0] #find first non empty value in the colum
    if column=="user_mentions" or column=="hashtags":
        value=tweets[tweets[column].apply(lambda c: c!=[])][column].iloc[0] # for column type "list"
    print("Column name: {colname} \nColumn type: {coltype} \nColumn data(example): {coldata} \n\n"
          .format(colname=column, coltype=type(value),coldata=value))

#### So we have the following list of columns:
 - **coordinates** - user coordinates, can be None or Python type Dict(needs to be split to two columns - longitude and latitude)
 - **created_at**  - creation date, Python type String (needs to be converted to type Date + timezone needs to be changed from UTC to MST)
 - **extended_tweet** - tweet text, Python type String
 - **hashtags** - hashtags, Python type List of Dicts(needs to be converted to String)
 - **lang** - language , Python type String
 - **name** - twitter account name, Python type String
 - **screen_name** - screen name, Python type String
 - **text** - truncated tweet text, Python type String
 - **user_location** - location, Python type String
 - **user_mentions** - user mentions, Python type List of Dicts(needs to be converted to String)

tweets_cleaned = tweets #create a copy of tweets dataframe, will do cleaning using this copy

#### Convert column `created_at`  from String to Date and change timezone from UTC to MST

print(tweets_cleaned[tweets_cleaned["created_at"].notnull()]["created_at"].head()) #print first non-empty row for "created_at" column

tweets_cleaned["created_at_date"]= pd.to_datetime(tweets_cleaned["created_at"]) ## convert to timestamp
tweets_cleaned["created_at_date"]=tweets_cleaned["created_at_date"].dt.tz_localize('UTC').dt.tz_convert('MST')

print(tweets_cleaned[tweets_cleaned["created_at_date"].notnull()]["created_at_date"].head())  ## Print the transformed data

#### Convert column `coordinates` from dictionary to two columns `latitude` and `longitude`

print(tweets_cleaned[tweets_cleaned["coordinates"].notnull()]["coordinates"].head()) ## Print the original data

def convertCoordToList(v):  ## helper function
    if v is None:
        v=[]
    else:
        v= v['coordinates']
    return v

tweets_cleaned['coordinates_list']= tweets_cleaned['coordinates'].map(convertCoordToList) #Convert coordinates from Dict to List

#Split coordinates_list column into two: latitude and longitude
tweets_cleaned[['latitude','longitude']] = pd.DataFrame(tweets_cleaned['coordinates_list'].values.tolist())

print(tweets_cleaned[tweets_cleaned["coordinates"].notnull()][['longitude','latitude']].head()) ## Print the transformed data

#### Convert column `hashtags`  from dictionary to strings
If there were no hashtags in a cell, we will fill that cell in our dataframe with a single space `' '`. 

print(tweets_cleaned[tweets_cleaned["hashtags"].apply(lambda c: c!=[])]["hashtags"].head()) ## Print the original data

def convertHashToString(v): ## helper function
    if v == []:
        v=' '
    else:
        v=' '.join([d['text'] for d in v])
    return v

tweets_cleaned['hashtags_string']= tweets_cleaned['hashtags'].map(convertHashToString) ## Do the conversion

print(tweets_cleaned[tweets_cleaned["hashtags_string"]!=' ']["hashtags_string"].head()) ## Print the transformed data

#### Convert column `user_mentions` from dictionary to string

We don't want the `user_mentions` column to be in a dictionary format as it is easier to work with as a string. Let's convert it.

The following is a sample of what the data looks like.

print(tweets_cleaned[tweets_cleaned["user_mentions"].apply(lambda c: c!=[])]["user_mentions"].head()) 

**Exercise2:** Using the `hashtags` conversion above as an example, complete the following cell in order to convert the `user_mentions` column to strings. 

def convertUserToString(v):
    if v == []:
        v=' '
    else:
        v=' '.join([d['screen_name'] for d in v])
    return v

###### Enter your code below

######

#### Check your answer
Only screen_name's should be extracted from the distionary.

print(tweets_cleaned[tweets_cleaned["user_string"]!=' ']["user_string"].head()) 

## Step 3. Cleaning the data

Some data cleanup was done in step 2. Let's take a closer look at the `extended_tweet` column, since we will use it in further during our data analysis. 

First, let's look at what some of the tweets look like. We'll set an output parameter and then access the first few tweets. 

pd.set_option('max_colwidth', 700) ## Setting col_widht wider - so we can see the entire tweet
tweets_cleaned['extended_tweet'].head(10)

####  Remove URLS 
There are lots of links in this column that might make our future data analysis messy and likely do not have any meaning. Let's remove them together with any links starting with www.  

We will accomplish this using regular expression: **'http\S+|www.\S+'** (Note that **\S+** means one or more non-whitespace character)

tweets_cleaned['extended_tweet']=tweets_cleaned['extended_tweet'].str.replace('http\S+|www.\S+', '', case=False)

#### Exercise3: remove '\n' (new line symbol) from 'extended_tweet' column

Another item that might interfere with our analysis are new line characters. Let's remove these as well in the cell below.  
(Hint: regular expression in this case will be just one character **'\n'**)

##### Enter your code below


######

#### Check your answer
Print the first 10 tweets and see if \n is removed

tweets_cleaned['extended_tweet'].head(10)

#### Remove user mentions and hashtags from the 'extended_tweet' column and save it as 'extended_tweet_cleaned' column
There are separate columns for user mentions and hashtags. We might need to analyse tweet text without any hashtags/user mentions. However, we might also want to do some analysis on the tweets including the mentions and hashtags. As such, tweets with these items removed will be stored in a separate column. 

tweets_cleaned['extended_tweet_cleaned']=tweets_cleaned['extended_tweet'].str.replace('#', '',case=False)
tweets_cleaned['extended_tweet_cleaned'].head(5)

We will keep emoji, punctuation,  numbers and special characters for now. They might be useful for the analysis. We can always do more cleaning in the later notebooks

#### Let's take a look at the language column

tweets_cleaned["lang"].unique()

Column lang has only one value en so it can be safely ignored.

## Step 6. Publishing - saving data for future work

In order to work with this data further, let's save a new DataFrame. 

We'll simply select the desired columns from the working dataset into the output dataset and take a look at the output. Does the output match what we expected? How do the `extended_tweet` and `extended_tweet_clean` columns compare? 

Further below, we'll save the data to a csv file. 

tweets_cleaned

tweets_cleaned_final=tweets_cleaned[['created_at_date','hashtags_string','user_string','user_location',
                     'longitude','latitude','name', 'screen_name','extended_tweet','extended_tweet_cleaned']]
tweets_cleaned_final.head()

tweets_cleaned_final.to_csv('alberta_tweets_cleaned_feb_25.csv',index=False)


##  Conclusion

We've taken a preliminary look at some of the raw Twitter data that's been collected in order to further work with it. This included restructuring and reorganizing it and cleaning up the columns that we anticipate needing the most for our analysis.

Take-aways: 
* We've started working with some very popular libraries for doing data science in Python: pandas and matplotlib
* Examining, wrangling, and cleaning your data are important work that will take up large portions of time in the data science workflow
* This is an iterative process. We took a stab at what we need to clean up and refine now, but may very well find things we've missed as we dig deeper into the data. 

Next, we will get into data exploration further. 

## Solutions

#### Exercise1:

print("Tweet as seen in the **extended_tweet** column:\n")
###### Enter your code in place of None below
extended = tweets["extended_tweet"].iloc[18960]
######
print(str(extended) + "\n")

print("Tweet as seen in the **text** column:\n")
###### Enter your code in place of None below
text_tweet = tweets["text"].iloc[18960]
######
print(text_tweet)

#### Exercise2:

tweets_cleaned['user_string']=tweets_cleaned["user_mentions"].map(convertUserToString)

#### Exercise3:

tweets_cleaned['extended_tweet']=tweets_cleaned['extended_tweet'].str.replace('\n', '')
