==> 4.User_Info_Gather.ipynb <==
# Gathering Users Followers

In this notebook, we demonstrate how you may gather user's information using the Twitter API as wrapped by Tweepy.  Note however that due to twitter API limitations, we only gather up to 250 followers for each user, representing only a partial social network.

### Things you'll need to run this notebook
1. Your own Twitter developer credentials available which you can sign up for [here](https://developer.twitter.com/en/apply-for-access). Once you get your API keys, you can fill in the `YOUR CREDENTIALS HERE` section of the code below. It is however, not required for this workshop as we've already captured the data for you. 
2. A list of users you've identified in your topic analysis notebook

### What this notebook does

1. Connects to the twitter API in the first cell 
2. Iterates over the users you've identified in the topic model stored in `user_list` and gathers up to 250 of their followers. 
3. Saves that information as a JSON (JavaScript Object Notation) file 



import json
import time
from IPython.display import clear_output

try : 
    import tweepy
    from tweepy import Stream
    from tweepy import OAuthHandler
    from tweepy.streaming import StreamListener
except ImportError:
    !pip install tweepy --user
    from importlib import reload
    import site
    reload(site)
    import tweepy
    from tweepy import Stream
    from tweepy import OAuthHandler
    from tweepy.streaming import StreamListener

# Connect to twitter API


auth = tweepy.OAuthHandler('YOUR CREDENTIALS HERE', 
                           'YOUR CREDENTIALS HERE')
auth.set_access_token('YOUR CREDENTIALS HERE', 
                      'YOUR CREDENTIALS HERE')


api = tweepy.API(auth, wait_on_rate_limit=True, wait_on_rate_limit_notify=True, compression=True)

### Test if the api is working
user = api.me()
print (user.name)

# Here you would fill your list with the names of people you've identified in your topic analysis 
user_list=[]

followers_dict = {}
N = 0
for i,follower in enumerate(user_list):
    user = follower
    clear_output()
    print(user, i)
    follower_followers = []
    try:
        curse = tweepy.Cursor(api.followers, screen_name=user).items(250)
    except tweepy.TweepError:
        continue
    try:
        for page in curse:
            follower_followers.append(page._json)
    except tweepy.TweepError:
        continue

    followers_dict[user] = follower_followers
    for j in range(61):
        clear_output(wait=True)
        time.sleep(1)
        print("Waiting", 60 - j, "Seconds before proceeding")
        print(user, i)
#print(followers_dict)
# Save results to file 
with open('topic_users.json', 'w') as fp:
     json.dump(followers_dict, fp)
